requirements.txt
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
config==0.5.1
configparser==5.2.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
executing==0.8.2
fonttools==4.29.1
idna==3.3
iexfinance==0.5.0
ipykernel==6.6.1
ipython==8.0.0
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
lxml==4.7.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
multitasking==0.0.10
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathlib==1.0.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
postgres==4.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
psycopg2-pool==1.1
ptyprocess==0.7.0
pure-eval==0.2.1
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytickersymbols==1.8.53
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
stack-data==0.1.3
tomli==1.2.3
tornado==6.1
-e git+git@github.com:Michiel-DK/tracker.git@483f3f999273e9540a1cdd5464acb2f9908904d9#egg=tracker
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
yfinance==0.1.69
altair==4.2.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
black==21.12b0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
config==0.5.1
configparser==5.2.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.3
executing==0.8.2
fonttools==4.29.1
gitdb==4.0.9
GitPython==3.1.26
idna==3.3
iexfinance==0.5.0
importlib-resources==5.4.0
ipykernel==6.6.1
ipython==8.0.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multitasking==0.0.10
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
postgres==4.0
prometheus-client==0.13.1
prompt-toolkit==3.0.24
protobuf==3.19.4
psycopg2-binary==2.9.3
psycopg2-pool==1.1
ptyprocess==0.7.0
pure-eval==0.2.1
pyarrow==7.0.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytickersymbols==1.8.53
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
stack-data==0.1.3
streamlit==1.5.0
terminado==0.13.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.1
-e git+git@github.com:Michiel-DK/tracker.git@ebd9c14cf12a0c9324299e6f699a1a8a4e5b5e83#egg=tracker
traitlets==5.1.1
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
yfinance==0.1.69
zipp==3.7.0
